pyproject.toml
[project]
name = "testtaskbot-ascrru"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "loguru>=0.7.3",
    "python-decouple>=3.8",
]
